+++ src/datatypes/light.h
@@ -6,7 +6,7 @@
 /*   By: joaoteix <joaoteix@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/24 00:46:12 by joaoteix          #+#    #+#             */
-/*   Updated: 2024/04/24 09:47:55 by joaoteix         ###   ########.fr       */
+/*   Updated: 2024/05/05 16:41:11 by joaoteix         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ typedef struct s_light
 {
 	t_node3d	node;
 	t_color3	color;
-	float		f;
+	double		f;
 }	t_light;
 
 t_light	*light_new(void);
