+++ src/utils.h
@@ -6,7 +6,7 @@
 /*   By: jcat <joaoteix@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/31 15:17:10 by jcat              #+#    #+#             */
-/*   Updated: 2024/04/24 17:16:25 by joaoteix         ###   ########.fr       */
+/*   Updated: 2024/05/05 16:53:29 by joaoteix         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,14 +23,14 @@ t_vec3	perf_ray(t_vec3 *ldir, t_vec3 *normal);
 
 void	print_err(char *msg);
 
-float	fsign(float f);
+double	fsign(double f);
 
 int		tern(int cond, int a, int b);
 
-float	ftern(int cond, float a, float b);
+double	ftern(int cond, double a, double b);
 
 int		clamp(int a, int min, int max);
-float	fclamp(float a, float min, float max);
+double	fclamp(double a, double min, double max);
 void	arg_val(int argc, char **argv);
 
 void	ft_strjoin_morph(char **dst, char *other);
